Allow mpie bit to be writeable in mstatus
This bit is currently not r/w but this means that it is
not possible to change the mstatus mie state from within
an exception handler.  The mie bit is always set to its
previous state when mret is called.  This is an issue
with an RTOS where ecall may be called from a new task creation
which is in a critial section.  This mean that interrupts
including the system tick will not be triggered.

This functionality was tested with the code in this GitHub
gist https://gist.github.com/btashton/4aa7ed42bc81ff4716821636997d9df9

Signed-off-by: Brennan Ashton <bashton@brennanashton.com>